re-arranged script to reduce unrelated imports when running scripts
