--- goodman_pipeline/scripts/__init__.py
@@ -1,5 +1,0 @@
-from .scripts import (redccd,
-                      redspec,
-                      create_reference_lamp,
-                      redastrometry,
-                      redphotometry)  # pragma: no cover
